function template demonstration
The end goal of this program is to demonstrate an appropriate use for
function templates. I (the programmer) have been provided with a text
file containing both integer and string values. This program will read
from that input file and store the integer values in an integer vector
and the string values into a string vector. Using function templates
that will work with multiple data types, this program will output the
minimum and maximum values stored in both the integer and string
vectors. One function template will return the minimum value for both of
the vectors while a second function template will return the maximum
value for both of the vectors.